pusku

+++ controllers/blogs.js
@@ -1,23 +1,43 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
-const formatBlog = (blog) => {
-  return {
-    id: blog._id,
-    title: blog.title,
-    author: blog.author,
-    url: blog.url,
-    likes: blog.likes
+const getTokenFrom = (request) => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
   }
+  return null
 }
 
 blogsRouter.get('/', async(request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs.map(formatBlog))
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
+
+  response.json(blogs.map(Blog.format))
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
+  const body = request.body
+
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    console.log(decodedToken)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error : 'token missing or invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    console.log(blog)
+    console.log(user.id.toString())
+
+    if (blog.user.toString() !== user.id.toString()) {
+      return response.status(401).json({ error : 'unauthorized action'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
@@ -27,25 +47,36 @@ blogsRouter.delete('/:id', async(request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
+  const body = request.body
   try {
-    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
+    if (!decodedToken.id) {
+      return response.status(401).json({ error : 'token missing or invalid'})
+    }
     if (body.title === undefined) {
-      return response.status(400).json({error: 'content missing'})
+      return response.status(400).json({error: 'title missing'})
     }
     if (body.url === undefined) {
       return response.status(400).json({error: 'url missing'})
     }
 
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes === undefined ? 0 : body.likes
+      likes: body.likes === undefined ? 0 : body.likes,
+      user: user.id
     })
 
     const savedBlog = await blog.save()
-    response.json(formatBlog(savedBlog));
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    response.json(Blog.format(savedBlog));
   } catch (exception) {
     console.log(exception)
     response.status(500).json({error: 'something went wrong...'})
@@ -60,15 +91,13 @@ blogsRouter.put('/:id', async(request, response) => {
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes
+      likes: body.likes === undefined ? 0 : body.likes
     }
 
-    const savedBlog = await blog.save()
-
     Blog
-      .findByIdAndUpdate(request.params.id, savedBlog, { new: true })
+      .findByIdAndUpdate(request.params.id, blog, { new: true })
       .then(updatedBlog => {
-        response.json(formatBlog(updatedBlog))
+        response.json(Blog.format(blog))
       })
       .catch(error => {
         console.log(error)
